fix migration -id

+++ database/migrations/2022_07_29_081409_roles_permissions_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
     public function up()
     {
         Schema::create('roles_permission', function (Blueprint $table) {
-            $table->unsignedBigInteger('permission_id');
-            $table->unsignedBigInteger('role_id');
+            $table->unsignedInteger('permission_id');
+            $table->unsignedInteger('role_id');
             $table->timestamps();
 
             $table->foreign('permission_id')->references('id')->on('permissions')
